Refactor GUI: remove color theme system and Reference-page remnants
- Remove dynamic color theme (ajisai-theme.js / ajisai-config.js) from index.html;
  replace with static CSS custom properties in app-interface.css :root block
- Strip reference-page-only styles from ajisai-base.css (side-nav, content-wrapper,
  section, article, table, feature-grid, word-item, code/pre blocks, etc.);
  file reduced from ~783 lines to ~285 lines
- Extract shared interpreter execution utilities into interpreter-execution-utils.ts:
  collectUserWords, createExecutionSnapshot, syncInterpreterState,
  resolveExecutionException
- Remove duplicate implementations of those utilities from execution-controller.ts
  and step-executor.ts; both now import from the shared module
