+++ index.html
@@ -16,22 +16,6 @@
     <title>Ajisai</title>
 
 
-    <script src="/ajisai-config.js?v=20260408"></script>
-    <script src="/ajisai-theme.js?v=20260408"></script>
-    <style id="theme-vars"></style>
-    <script>
-
-        if (typeof AjisaiTheme !== 'undefined') {
-            const vars = AjisaiTheme.getAll();
-            let css = ':root {';
-            for (const [key, value] of Object.entries(vars)) {
-                css += `${key}: ${value};`;
-            }
-            css += '}';
-            document.getElementById('theme-vars').textContent = css;
-        }
-    </script>
-
     <link rel="stylesheet" href="public/ajisai-base.css?v=20260408">
     <link rel="stylesheet" href="app-interface.css?v=20260408">
 </head>
