+++ public/ajisai-base.css
@@ -1,13 +1,4 @@
 
-
-
-
-
-
-
-
-
-
 * {
     margin: 0;
     padding: 0;
@@ -191,19 +182,6 @@ header h1 a {
     color: var(--header-text-secondary);
 }
 
-address {
-    font-style: normal;
-    font-size: 0.875rem;
-    margin-top: 5px;
-    line-height: 1.4;
-    color: var(--header-text-tertiary);
-}
-
-address a {
-    margin-right: 10px;
-    color: var(--header-text-secondary);
-}
-
 
 
 
@@ -227,158 +205,6 @@ footer a {
 
 footer a:hover {
     color: var(--header-text);
-}
-
-
-
-
-.side-nav {
-    padding: 10px;
-    flex: var(--col-left-flex);
-    background: var(--gradient-parent);
-}
-
-.side-nav .nav-section {
-    line-height: 1.4;
-    margin-bottom: 20px;
-    padding: 10px;
-    word-break: break-all;
-    border: var(--border-main);
-    border-radius: var(--radius-main);
-    background-color: rgba(255, 255, 255, 0.5);
-}
-
-.side-nav ul {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-
-.side-nav li {
-    margin: 5px 0;
-}
-
-.side-nav .nav-section-title {
-    font-weight: bold;
-    margin: 10px 0;
-    padding-bottom: 5px;
-    border-bottom: 1px solid var(--color-secondary);
-}
-
-.side-nav .nav-section-title span {
-    font-size: 60%;
-    display: block;
-    color: var(--color-text-light);
-}
-
-.social-links {
-    list-style: none;
-    padding: 0;
-    display: flex;
-    flex-wrap: wrap;
-    gap: 5px;
-}
-
-.social-links li {
-    margin: 0;
-}
-
-.social-links a {
-    padding: 5px 10px;
-    color: var(--color-primary);
-    border-radius: var(--radius-main);
-    border: var(--border-main);
-    display: inline-block;
-    background: rgba(255, 255, 255, 0.5);
-}
-
-.social-links a:hover {
-    background: var(--color-light);
-}
-
-
-
-
-.content-wrapper {
-    display: flex;
-    flex: 1;
-    gap: 0;
-    margin: 0;
-    padding: 0;
-    background: var(--gradient-parent);
-    border-top: var(--border-main);
-}
-
-.main-content {
-    padding: 10px;
-    flex: var(--col-right-flex);
-    background: var(--gradient-parent);
-    min-height: 500px;
-}
-
-
-
-
-.panel {
-    background: var(--gradient-parent);
-    padding: 1rem;
-    overflow-y: auto;
-}
-
-
-
-
-section {
-    line-height: 1.6;
-    overflow: visible;
-    margin-bottom: 20px;
-    padding: 15px;
-    word-break: break-all;
-    border: var(--border-main);
-    background: var(--gradient-child);
-    border-radius: var(--radius-main);
-}
-
-section h2 {
-    font-size: 1.125rem;
-    margin-bottom: 0.75rem;
-    color: var(--heading-h2);
-    font-weight: 500;
-}
-
-section h3,
-section h4 {
-    font-size: 1rem;
-    font-weight: bold;
-    margin: 20px 0 10px 0;
-    color: var(--heading-h3);
-}
-
-section h3:first-child {
-    margin-top: 0;
-}
-
-
-
-
-.word-info-display {
-    font-size: 0.85rem;
-    font-weight: normal;
-    color: var(--color-text-light, #666);
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    max-width: calc(100% - 120px);
-    flex: 1;
-    min-width: 0;
-}
-
-article h2 {
-    text-align: center;
-    color: var(--heading-h2);
-    border-bottom: 2px solid var(--color-secondary);
-    padding-bottom: 10px;
-    margin: 0 0 20px 0;
 }
 
 
@@ -470,182 +296,6 @@ article h2 {
 
 
 
-
-code {
-    font-family: var(--font-stack-mono);
-    font-size: 14px;
-    background: var(--code-inline-bg);
-    padding: 2px 6px;
-    border-radius: 3px;
-    border: 1px solid var(--color-medium);
-}
-
-pre {
-    background: var(--code-bg);
-    color: var(--code-text);
-    border: 1px solid var(--color-secondary);
-    border-radius: var(--radius-main);
-    padding: 16px;
-    overflow-x: auto;
-    margin: 16px 0;
-    line-height: 1.5;
-}
-
-pre code {
-    background: none;
-    border: none;
-    padding: 0;
-    color: var(--code-text);
-    font-size: 13px;
-}
-
-pre code .comment {
-    color: var(--code-comment);
-}
-
-
-
-
-table {
-    line-height: 1.4;
-    width: 100%;
-    table-layout: fixed;
-    border-collapse: collapse;
-    word-break: break-all;
-    margin-top: 15px;
-}
-
-table th {
-    width: 30%;
-    padding: 10px;
-    vertical-align: middle;
-    background-color: color-mix(in srgb, var(--color-primary) 10%, transparent);
-    border: solid 0.5px var(--color-secondary);
-}
-
-table td {
-    padding: 10px;
-    vertical-align: middle;
-    background-color: transparent;
-    border: solid 0.5px var(--color-secondary);
-}
-
-
-
-
-.feature-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-    gap: 15px;
-    margin: 20px 0;
-}
-
-.feature {
-    text-align: center;
-    padding: 20px;
-    background: rgba(255, 255, 255, 0.5);
-    border-radius: var(--radius-main);
-    border: var(--border-main);
-}
-
-.feature-icon {
-    font-size: 48px;
-    margin-bottom: 10px;
-}
-
-.feature h4 {
-    border-bottom: none;
-    margin: 10px 0;
-}
-
-.card-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-    gap: 15px;
-    margin: 20px 0;
-}
-
-.card {
-    background: rgba(255, 255, 255, 0.7);
-    border: var(--border-main);
-    border-radius: var(--radius-main);
-    padding: 20px;
-    transition: box-shadow 0.2s;
-}
-
-.card:hover {
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-}
-
-.card h4 {
-    margin-top: 0;
-    color: var(--color-primary);
-}
-
-
-
-
-.word-item {
-    background: rgba(255, 255, 255, 0.7);
-    border: 1px solid var(--color-secondary);
-    border-radius: var(--radius-main);
-    padding: 16px;
-    margin: 12px 0;
-}
-
-.word-item h4 {
-    margin: 0 0 8px;
-    color: var(--color-primary);
-    font-family: var(--font-stack-mono);
-    font-size: 18px;
-    border-bottom: none;
-}
-
-.word-item p {
-    margin: 8px 0;
-}
-
-.word-example {
-    background: var(--code-bg);
-    border-radius: 4px;
-    padding: 12px;
-    margin-top: 8px;
-}
-
-.word-example code {
-    display: block;
-    white-space: pre-wrap;
-    color: var(--code-text);
-    background: none;
-    border: none;
-}
-
-
-
-
-ul,
-ol {
-    margin: 0 0 16px 0;
-    padding-left: 24px;
-}
-
-li {
-    margin-bottom: 8px;
-}
-
-
-
-
-small {
-    font-size: 80%;
-    color: var(--color-secondary);
-    display: block;
-    margin-top: 5px;
-}
-
-
-
-
 .display-area,
 .state-display,
 .words-display {
@@ -726,57 +376,8 @@ small {
     }
 
 
-    .content-wrapper {
-        flex-wrap: wrap-reverse;
-        gap: 0;
-    }
-
-    .side-nav,
-    .main-content {
-        flex: auto;
-        width: 100%;
-        padding: 10px 0;
-    }
-
-    .side-nav {
-        background: none;
-    }
-
-    .side-nav .nav-section {
-        padding: 10px;
-        margin: 0 10px 20px 10px;
-    }
-
-    .main-content {
-        padding: 10px 0;
-        min-height: auto;
-    }
-
-    .main-content section {
-        padding: 10px;
-        margin: 10px;
-    }
-
-
-    .feature-grid,
-    .card-grid {
-        grid-template-columns: 1fr;
-    }
-
-
-    pre {
-        font-size: 12px;
-        padding: 10px;
-    }
-
-
     #run-btn {
         padding: 0.75rem 1rem;
         font-size: 1rem;
-    }
-
-    .word-info-display {
-        max-width: 100%;
-        font-size: 0.8rem;
     }
 }
